+++ src/generalstring.v
@@ -84,11 +84,11 @@ pub fn GeneralString.parse(mut p Parser) !GeneralString {
 }
 
 // decode tries to decode bytes array into GeneralString or return error on fails.
-pub fn GeneralString.decode(src []u8) !(GeneralString, i64) {
+pub fn GeneralString.decode(src []u8) !(GeneralString, int) {
 	return GeneralString.decode_with_rule(src, .der)!
 }
 
-fn GeneralString.decode_with_rule(bytes []u8, rule EncodingRule) !(GeneralString, i64) {
+fn GeneralString.decode_with_rule(bytes []u8, rule EncodingRule) !(GeneralString, int) {
 	tag, length_pos := Tag.decode_with_rule(bytes, 0, rule)!
 	if !tag.equal(default_generalstring_tag) {
 		return error('Unexpected non-generalstring tag')
