+++ src/null.v
@@ -47,7 +47,7 @@ pub fn Null.parse(mut p Parser) !Null {
 }
 
 // Null.decode read Null from bytes.
-pub fn Null.decode(bytes []u8) !(Null, i64) {
+pub fn Null.decode(bytes []u8) !(Null, int) {
 	tag, length_pos := Tag.decode(bytes)!
 	if !tag.equal(default_null_tag) {
 		return error('Null: get unexpected tag')
